first draft make test-src working
