+++ transforms/language/html2parquet/Dockerfile.python
@@ -19,20 +19,9 @@ RUN  pip install data-processing-dist/${DPK_WHEEL_FILE_NAME}
 
 # END OF STEPS destined for a data-prep-kit base image 
 
-COPY --chown=dpk:root src/ src/
-COPY --chown=dpk:root pyproject.toml pyproject.toml 
+COPY --chown=dpk:root dpk_html2parquet/ dpk_html2parquet/
 COPY --chown=dpk:root requirements.txt requirements.txt
-RUN pip install --no-cache-dir -e .
-
-# copy transform main() entry point to the image 
-COPY ./src/html2parquet_transform_python.py .
-
-# copy some of the samples in
-COPY ./src/html2parquet_local.py local/
-
-# copy test
-COPY test/ test/
-COPY test-data/ test-data/
+RUN pip install -r requirements.txt
 
 # Set environment
 ENV PYTHONPATH /home/dpk
