+++ transforms/language/html2parquet/dpk_html2parquet/transform_python.py
@@ -10,14 +10,15 @@
 # limitations under the License.
 ################################################################################
 
-import time
+import sys
 
 from data_processing.runtime.pure_python import PythonTransformLauncher
 from data_processing.runtime.pure_python.runtime_configuration import (
     PythonTransformRuntimeConfiguration,
 )
 from data_processing.utils import get_logger
-from html2parquet_transform import Html2ParquetTransformConfiguration
+from dpk_html2parquet.transform import Html2ParquetTransformConfiguration
+from data_processing.utils import ParamsUtils
 
 
 logger = get_logger(__name__)
@@ -35,6 +36,24 @@ class Html2ParquetPythonTransformConfiguration(PythonTransformRuntimeConfigurati
         """
         super().__init__(transform_config=Html2ParquetTransformConfiguration())
 
+
+
+#Class used by the notebooks to ingest binary files and create parquet files
+class Html2ParquetRuntime():
+    def __init__(self, **kwargs):
+        self.params={}
+        for key, value in kwargs:
+            self.params[key]=value
+
+    
+    def ingest(self):
+        sys.argv = ParamsUtils.dict_to_req(d=(self.params))
+        # create launcher
+        launcher = PythonTransformLauncher(Html2ParquetPythonTransformConfiguration())
+        # launch
+        return_code = launcher.launch()
+        return return_code
+    
 
 if __name__ == "__main__":
     launcher = PythonTransformLauncher(Html2ParquetPythonTransformConfiguration())
