--- transforms/language/html2parquet/ray/requirements.txt
@@ -1,3 +1,0 @@
-dpk-html2parquet-transform-python==0.2.2.dev2
-data-prep-toolkit[ray]==0.2.2.dev2
-trafilatura==1.12.0
