--- transforms/language/html2parquet/ray/test-data/expected/metadata.json
@@ -1,54 +1,0 @@
-{
-  "pipeline": "pipeline_id",
-  "job details": {
-    "job category": "preprocessing",
-    "job name": "html2parquet",
-    "job type": "ray",
-    "job id": "job_id",
-    "start_time": "2024-10-15 11:13:06",
-    "end_time": "2024-10-15 11:13:07",
-    "status": "success"
-  },
-  "code": null,
-  "job_input_params": {
-    "output_format": "markdown",
-    "favor_precision": "True",
-    "favor_recall": "True",
-    "checkpointing": false,
-    "max_files": -1,
-    "random_samples": -1,
-    "files_to_use": [
-      ".html",
-      ".zip"
-    ],
-    "number of workers": 1,
-    "worker options": {
-      "num_cpus": 0.8,
-      "max_restarts": -1
-    },
-    "actor creation delay": 0
-  },
-  "execution_stats": {
-    "cpus": 10,
-    "gpus": 0,
-    "memory": 38.15873718261719,
-    "object_store": 2.0,
-    "execution time, min": 0.008651264508565267
-  },
-  "job_output_stats": {
-    "source_files": 2,
-    "source_size": 460391,
-    "result_files": 2,
-    "result_size": 16898,
-    "processing_time": 0.16381311416625977,
-    "nrows": 3
-  },
-  "source": {
-    "name": "/Users/sungeunan/Desktop/temp/data-prep-kit/transforms/language/html2parquet/ray/test-data/input",
-    "type": "path"
-  },
-  "target": {
-    "name": "/Users/sungeunan/Desktop/temp/html2parquet-output",
-    "type": "path"
-  }
-}
