+++ transforms/language/html2parquet/test/test_html2parquet.py
@@ -11,12 +11,11 @@
 ################################################################################
 
 import os
-import pyarrow as pa
 from data_processing.data_access.data_access_local import DataAccessLocal
 from data_processing.test_support import get_files_in_folder
 from data_processing.test_support.transform import AbstractBinaryTransformTest
 from data_processing.utils import TransformUtils
-from html2parquet_transform import Html2ParquetTransform
+from dpk_html2parquet.transform import Html2ParquetTransform
 
 
 class TestHtml2ParquetTransform(AbstractBinaryTransformTest):
