+++ transforms/language/html2parquet/test/test_html2parquet_python.py
@@ -13,14 +13,13 @@
 import ast
 import os
 
+from data_processing.runtime.pure_python import PythonTransformLauncher
 from data_processing.test_support.launch.transform_test import (
     AbstractTransformLauncherTest,
 )
-from data_processing_ray.runtime.ray import RayTransformLauncher
-from html2parquet_transform_ray import Html2ParquetRayTransformConfiguration
+from dpk_html2parquet.transform_python import Html2ParquetPythonTransformConfiguration
 
-
-class TestRayHtml2ParquetTransform(AbstractTransformLauncherTest):
+class TestPythonHtml2ParquetTransform(AbstractTransformLauncherTest):
     """
     Extends the super-class to define the test data for the tests defined there.
     The name of this class MUST begin with the word Test so that pytest recognizes it as a test class.
@@ -31,19 +30,22 @@ class TestRayHtml2ParquetTransform(AbstractTransformLauncherTest):
         basedir = os.path.abspath(os.path.join(os.path.dirname(__file__), basedir))
         config = {
             "data_files_to_use": ast.literal_eval("['.html','.zip']"),
-            "run_locally": True,
+            "html2parquet_output_format": "markdown",
         }
+        # this is added as a fixture to remove these columns from comparison
+        ignore_columns = ["date_acquired"]
+
         fixtures = []
-        launcher = RayTransformLauncher(Html2ParquetRayTransformConfiguration())
+        launcher = PythonTransformLauncher(Html2ParquetPythonTransformConfiguration())
         fixtures.append(
             (
                 launcher,
                 config,
                 basedir + "/input",
                 basedir + "/expected",
-                # this is added as a fixture to remove these columns from comparison
-                ["date_acquired"],
+                ignore_columns,
 
             )
         )
         return fixtures
+
